db: replace FlushTimeLogger with generic OpLogger
Now that we want to log compaction lengths, make the logger generic
and also use it for compactions. Also buffer writes in memory and
periodically flush. This avoids the overhead of going to disk on
every write, doesn't require a resizing list, and still writes to
disk in a timely fashion. There is one thread that is responsible
for flushing the logs across all loggers.