all: plumb req start times through
In order to accurately measure service times,
we need to start from when requests entered Cassandra.
This requires a lot of plumbing (TLS, no thanks).
We can now successfully record read service times in cassandra,
but we are counting more than expected.
Why is this?

Are we double counting requests due to strange behavior in C*,
or are we measuring internal requests?
TODO: figure this out if it affects latency distributions.